@@ -1,5 +1,8 @@
 {
   "name": "mmm",
+  "jest": {
+    "coverageReporters": []
+  },
   "homepage": ".",
   "version": "0.1.0",
   "private": true,
@@ -8,26 +11,31 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "d3": "^5.15.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
     "file-saver": "^2.0.2",
     "font-awesome": "^4.7.0",
+    "jest-enzyme": "^7.1.2",
     "mdbootstrap": "^4.19.0",
     "node-sass": "^4.14.1",
     "papaparse": "^5.2.0",
-    "react": "^16.12.0",
+    "react": "^16.14.0",
     "react-addons-update": "^15.6.2",
-    "react-dom": "^16.12.0",
+    "react-dom": "^16.14.0",
     "react-modal": "^3.11.2",
-    "react-redux": "^7.2.0",
+    "react-redux": "^7.2.2",
     "react-scripts": "3.4.3",
+    "react-test-renderer": "^16.14.0",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "scss": "^0.2.4",
     "semantic-ui-react": "^0.88.2"
   },
   "scripts": {
+    "__testComment__": "watchAll disables watch on completion. json and outputfile pipe results to testOutput",
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "react-scripts test \"--watchAll=false\" \"--silent\" \"--reporters=jest-silent-reporter\"",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -44,5 +52,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "jest-silent-reporter": "^0.2.1"
   }
 }
